+++ core/src/test/java/feign/FeignTest.java
@@ -106,7 +106,7 @@ public class FeignTest {
     @RequestLine("POST /")
     Observable<Response> observableResponse();
 
-    @dagger.Module(overrides = true, library = true)
+    @dagger.Module(library = true)
     static class Module {
       @Provides(type = SET)
       Encoder defaultEncoder() {
@@ -127,19 +127,6 @@ public class FeignTest {
           }
         };
       }
-
-      // just run synchronously
-      @Provides
-      @Singleton
-      @Named("http")
-      Executor httpExecutor() {
-        return new Executor() {
-          @Override
-          public void execute(Runnable command) {
-            command.run();
-          }
-        };
-      }
     }
   }
 
@@ -154,7 +141,8 @@ public class FeignTest {
           Feign.create(
               TestInterface.class,
               "http://localhost:" + server.getPort(),
-              new TestInterface.Module());
+              new TestInterface.Module(),
+              new RunSynchronous());
 
       final AtomicBoolean success = new AtomicBoolean();
 
@@ -196,7 +184,8 @@ public class FeignTest {
           Feign.create(
               TestInterface.class,
               "http://localhost:" + server.getPort(),
-              new TestInterface.Module());
+              new TestInterface.Module(),
+              new RunSynchronous());
 
       final AtomicBoolean success = new AtomicBoolean();
 
@@ -227,6 +216,21 @@ public class FeignTest {
     }
   }
 
+  @Module(library = true, overrides = true)
+  static class RunSynchronous {
+    @Provides
+    @Singleton
+    @Named("http")
+    Executor httpExecutor() {
+      return new Executor() {
+        @Override
+        public void execute(Runnable command) {
+          command.run();
+        }
+      };
+    }
+  }
+
   @Test
   public void incrementString() throws IOException, InterruptedException {
     final MockWebServer server = new MockWebServer();
@@ -238,7 +242,8 @@ public class FeignTest {
           Feign.create(
               TestInterface.class,
               "http://localhost:" + server.getPort(),
-              new TestInterface.Module());
+              new TestInterface.Module(),
+              new RunSynchronous());
 
       final AtomicBoolean success = new AtomicBoolean();