+++ gson/src/main/java/feign/gson/GsonModule.java
@@ -42,7 +42,7 @@ import java.util.concurrent.atomic.AtomicBoolean;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-@dagger.Module(library = true, overrides = true)
+@dagger.Module(library = true)
 public final class GsonModule {
 
   @Provides(type = SET)